When "stepping" in the debugger, don't refresh immediately.
Instead, we refresh the debugger after 10 milliseconds, to avoid
flickering if the instruction or subroutine finishes very quickly.

This is definitely an improvement for single-stepping.  I'm not sure
if we should do the same for action_run(); in that case, if there is a
particular breakpoint that's being hit repeatedly, then without the
flicker, it might not be apparent that anything actually happened.


git-svn-id: svn://svn.code.sf.net/p/tilem/code/trunk@467 05cf5367-e1a8-405a-8a5c-7f956f11530a