Cleaning up the menu, removing stuff we don't need.
load-skin, hide-skin, toggle-speed: these are now in the prefs dialog;
no need to clutter the menu.

output-terminal: seriously, is there any reason to keep this?

hide-border: I don't really like this either - it goes against all the
rules of good interface design - but I know there are some people who
like this sort of thing, and I can see the argument that it saves a
little bit of space on your screen.  Well, I guess if there's really a
huge demand from users, we could conceivably bring this back, but (a)
the emulator window would need to provide its own ways of moving and
resizing the window, if the WM isn't doing so, and (b) it should be an
option in the prefs dialog, not in the menu.


git-svn-id: svn://svn.code.sf.net/p/tilem/code/trunk@472 05cf5367-e1a8-405a-8a5c-7f956f11530a