Don't refresh screen needlessly.
If we've already asked the GUI thread to update the screen, but it
hasn't done so yet, there's no point in adding another idle callback.
The old behavior was problematic when running at full speed, when the
core thread may be generating frames faster than the GUI thread is
able to handle them.


git-svn-id: svn://svn.code.sf.net/p/tilem/code/trunk@473 05cf5367-e1a8-405a-8a5c-7f956f11530a