Give priority to non-core-thread users of the calc mutex.
When running at full speed, the core needs to keep the calc mutex
locked almost all of the time.  In fact, there's only a very short
window during each iteration when the mutex is unlocked and then,
almost immediately, locked again.

In the case of glibc on Linux, if another thread is blocked in
pthread_mutex_lock() when we call pthread_mutex_unlock(), the blocked
thread will be woken up, but it won't receive ownership of the mutex
automatically; instead, the mutex is momentarily unlocked, and
whichever thread first claims it gets ownership.  On SMP systems, the
core thread will often win this race, so the other thread never gets a
chance to run.

(Other mutex implementations may, and probably do, have similar
problems.)

With this change, other threads that want to access the calculator
will first set a flag asking the core thread to yield the mutex
temporarily.  This is ugly and maybe not the best solution.

GUI routines, and any other routines, that access the calculator
directly must use tilem_calc_emulator_lock() and
tilem_calc_emulator_unlock() rather than accessing the mutex directly.


git-svn-id: svn://svn.code.sf.net/p/tilem/code/trunk@475 05cf5367-e1a8-405a-8a5c-7f956f11530a