New core loop / task system.
Essentially, this allows the GUI to schedule "tasks" to be run in the
core thread; if no task is running, emulation runs normally.  Task
functions have direct access to the calculator, including the ability
to run the CPU (using the new emucore API.)

This should make it possible to implement linking and macros in a more
sane way.  All of the old linking code is not working at this point.


git-svn-id: svn://svn.code.sf.net/p/tilem/code/trunk@493 05cf5367-e1a8-405a-8a5c-7f956f11530a