Improved handling of hardware exceptions.
(By exceptions, I'm referring to events that normally force the
calculator to reset.  These include executing in restricted memory
areas, and (on the Nspire) executing invalid opcodes.)

The former handling of these events was broken in most cases (a
function such as xp_z80_rdmem_m1 would try to reset the calculator
directly, but the CPU would then go on to increment the PC, skipping
the first byte of the boot code.)

In addition, in the future, I'd like for the debugger to be able to
catch these events, rather than simply resetting the emulator.  New
Z80 emulation flags have been added to make this possible.


git-svn-id: svn://svn.code.sf.net/p/tilem/code/trunk@513 05cf5367-e1a8-405a-8a5c-7f956f11530a