Track zoom factor rather than absolute window size.
This means that we try to preserve the same zoom factor when switching
between skin and no-skin modes, or from one skin to another.  Zoom
factor "1.0" is defined as the minimum size allowed (where each
row/column of the LCD corresponds to at least one row/column on the
screen.)

Also, preserve the skin's aspect ratio even if the WM doesn't.


git-svn-id: svn://svn.code.sf.net/p/tilem/code/trunk@525 05cf5367-e1a8-405a-8a5c-7f956f11530a