Fix file filters in save dialogs.
Old code was using the strings returned by tifiles_vartype2fext and
tifiles_fext_of_group, which use an uppercase '8X' to indicate the
83+.  This broke the file chooser, since prompt_save_file() expects
filter strings to be lowercase.

As a side effect of fixing this, the default filename now also uses a
lowercase 'x'.  I actually prefer this behavior.  But do other people
prefer the uppercase X?  We could do it either way.


git-svn-id: svn://svn.code.sf.net/p/tilem/code/trunk@539 05cf5367-e1a8-405a-8a5c-7f956f11530a