Add an option to switch between logical/absolute modes.
Move the memory view into its own file.

When switching modes, we preserve the current cursor position in the
memory view (as well as in the disassembly view.)  It would be good to
preserve the view position as well - i.e., the same bytes should be on
screen before and after switching modes, assuming there are no page
boundaries involved - but gtk_tree_view_scroll_to_cell() seems not to
work properly immediately after the model has been changed (?) and
there doesn't seem to be any way of setting the cursor that doesn't
also select and scroll to the given cell.


git-svn-id: svn://svn.code.sf.net/p/tilem/code/trunk@541 05cf5367-e1a8-405a-8a5c-7f956f11530a