Select model based on the filenames passed on the command line.
If files are of multiple types, we choose the "best" model to support
all of them, but we don't choose a model newer than the newest file
type.  For example, if an 82P file is passed, we only look for a TI-82
ROM (if it's an assembly program, it won't work on a TI-83 or TI-83
Plus.)  But if an 82P file and an 8XP file are passed at the same
time, we look for a TI-83 Plus ROM.

In general, we prefer models with more memory and features.  The order
of preferences is as follows:

 for .83? files:  TI-83 (or TI-82 Stats) > TI-76.fr

 for .8x? files:  TI-84+ SE > TI-84+ > TI-83+ SE > TI-83+ > Nspire-84+



git-svn-id: svn://svn.code.sf.net/p/tilem/code/trunk@553 05cf5367-e1a8-405a-8a5c-7f956f11530a