Wake up calculator correctly, especially after a reset.
Delay for 1 second before checking whether wake-up is needed, since
the calc takes some time to initialize itself (most of which, I
believe, is a deliberate delay to ensure the calc doesn't try to start
up without a good power supply.)

More generally, most task functions don't want to pause when the calc
is turned off; if we're just planning to wait for some amount of time
before pressing a key or sending a byte to the link port, the emulator
should continue running.  An additional arg to tilem_em_run()
indicates whether we should pause when the CPU power is turned off.


git-svn-id: svn://svn.code.sf.net/p/tilem/code/trunk@555 05cf5367-e1a8-405a-8a5c-7f956f11530a