Only return regular files from the file chooser.
If the user types the name of a directory followed by a slash, and
presses enter, the file chooser will switch to showing the contents of
that directory.  This is good.

If the user types the name of a directory not followed by a slash, and
presses enter, the file chooser returns, just as if the user had
selected a regular file.  This is quite annoying if the name of a
directory is also a prefix for one or more names of other files - in
which case, the user presses enter expecting to view the contents of
that directory, but instead (because the completion wasn't unique, and
therefore no slash was appended) the dialog simply closes.

I'm not sure what the intention of the GTK+ developers was.  Should we
perhaps, when the user selects a directory, call
gtk_file_chooser_set_current_folder?


git-svn-id: svn://svn.code.sf.net/p/tilem/code/trunk@583 05cf5367-e1a8-405a-8a5c-7f956f11530a