When reverting state, do it "in place".
Instead of creating a new calc, we can simply reload the state into
the existing calc.  This means that breakpoints will be preserved; it
also means that stuff like macros and screenshots will keep working.
We might want to change this behavior in the future, but the idea is
to keep the behavior of "revert" consistent with "reset".


git-svn-id: svn://svn.code.sf.net/p/tilem/code/trunk@603 05cf5367-e1a8-405a-8a5c-7f956f11530a