Work around a bug in GtkCellRendererCombo.
On Windows, combo boxes are displayed in "list style" by default
(since that is closer to the style of normal Windows programs.)  In
the case of GtkCellRendererCombo, this means even after the user has
changed the value, the cell stays in "editing mode" and does not emit
the "edited" signal.  The cell will not consider itself "edited" until
the user does something else to confirm the input, such as pressing
Enter or Tab.  This is seriously counterintuitive.

(This is related to GNOME bug #317387.)

The upshot of all this is that, in the "select a program slot" dialog
for the TI-81, if the user tries to change the program slot and then
*clicks OK* rather than pressing Enter, the dialog will ignore the
user's input.

(I also tried forcibly calling gtk_widget_child_focus() on the dialog
after gtk_dialog_run() - this workaround seems to work on X11 for
GCRCs with has-entry set - but for some reason this does not work on
Windows.  I still think a better workaround should be possible, maybe
using the editing-started signal, but it will be ugly.)

For now, I'm simply disabling list style for this particular case - it
may seem slightly weird and out of place on Windows, but at least it
should not be silently ignoring the user's input.


git-svn-id: svn://svn.code.sf.net/p/tilem/code/trunk@653 05cf5367-e1a8-405a-8a5c-7f956f11530a