Use new ticables2 raw API if possible.
This API will be added in libticables2 1.3.4 (not yet released) and
works substantially better than the old API for BlackLink and
ParallelLink.  You can still compile and run TilEm using an old
version of ticables2, but the option to connect a BlackLink/
ParallelLink cable will be disabled in that case.


git-svn-id: svn://svn.code.sf.net/p/tilem/code/trunk@692 05cf5367-e1a8-405a-8a5c-7f956f11530a