Hide external link options if libusb-win32 is missing.
The radio button is disabled, and in place of the usual options, a
warning message is displayed.

Checking for libusb is obviously not a foolproof way of determining
whether the necessary drivers are installed - the only way to be sure
of that is to try it.  But if libusb-win32 isn't found, then we can be
sure that TiLP isn't installed (which is the most likely scenario).
(If for some reason you've installed the tilp drivers for the older
link cables but *not* USB cables... well, you can always install
libusb0.dll manually, or just stick a copy in the tilem directory.
I'll assume that anyone interested in doing that already knows what
they're doing.  Expert users will also note that I haven't disabled
the command-line interface.)


git-svn-id: svn://svn.code.sf.net/p/tilem/code/trunk@693 05cf5367-e1a8-405a-8a5c-7f956f11530a