Add audio filter.
Similar to the grayscale filter, this is an auxiliary object that can
be attached to a TilemCalc, which monitors the calc state and produces
a stream of output.  The API is somewhat more complicated than
TilemGrayLCD, because it is designed to be usable with any of the wide
variety of sound hardware and APIs (as well as because, unlike video,
audio output does not tolerate any sloppiness in terms of timing.)
See audio.c for an overview of the API and implementation.

Audio backend and GUI are yet to be done (the filter has been tested
with some kludgy backend code that needs to be largely rewritten. :P)



git-svn-id: svn://svn.code.sf.net/p/tilem/code/trunk@698 05cf5367-e1a8-405a-8a5c-7f956f11530a